mod: Urls 파일 변경 및 적용
