+++ feature/areaverification/src/main/java/com/acon/acon/feature/areaverification/composable/PreferenceMapScreen.kt
@@ -52,7 +52,7 @@ fun PreferenceMapScreen(
 
     viewModel.collectSideEffect { effect ->
         when (effect) {
-            is AreaVerificationHomeSideEffect.NavigateToSystemLocationSettings -> {
+            is AreaVerificationSideEffect.NavigateToSystemLocationSettings -> {
                 val intent = Intent(Settings.ACTION_LOCATION_SOURCE_SETTINGS)
                 context.startActivity(intent)
             }
